sync

+++ database/migrations/2019_07_02_114259_create_orders_table.php
@@ -22,7 +22,7 @@ class CreateOrdersTable extends Migration
             $table->integer('car_id')->unsigned()->comment("汽车");
             $table->integer('status')->default(0)->comment("订单状态:客户未到店0、已到店1");
             $table->integer('pay_log_id')->unsigned()->comment("预支付订单id");
-            $table->integer('payment_count')->nullable()->comment("支付金额");
+            $table->float('payment_count')->nullable()->comment("支付金额");
             $table->integer('payment_status')->nullable()->comment("支付状态0,1");
             $table->json('info')->nullable()->comment("详细信息");
             $table->timestamp('created_at')->default(DB::raw('CURRENT_TIMESTAMP'));
